refactored test files to better structure
