+++ src/Codeception/Module/WebDriver.php
@@ -1,7 +1,6 @@
 <?php
 namespace Codeception\Module;
 
-use Codeception\Lib\TestDescriptor;
 use Codeception\Exception\ConnectionException;
 use Codeception\Exception\ElementNotFound;
 use Codeception\Exception\MalformedLocatorException;
@@ -20,6 +19,7 @@ use Codeception\PHPUnit\Constraint\Page as PageConstraint;
 use Codeception\PHPUnit\Constraint\WebDriver as WebDriverConstraint;
 use Codeception\PHPUnit\Constraint\WebDriverNot as WebDriverConstraintNot;
 use Codeception\TestCase;
+use Codeception\TestDescriptor;
 use Codeception\Util\Debug;
 use Codeception\Util\Locator;
 use Codeception\Util\Uri;
