+++ src/Codeception/Subscriber/Console.php
@@ -9,9 +9,9 @@ use Codeception\Events;
 use Codeception\Lib\Console\Message;
 use Codeception\Lib\Console\Output;
 use Codeception\Lib\Notification;
-use Codeception\Lib\Suite;
 use Codeception\Step;
 use Codeception\Step\Comment;
+use Codeception\Suite;
 use Codeception\TestCase;
 use Codeception\TestCase\Interfaces\ScenarioDriven;
 use Codeception\Util\Debug;
